WIP: linalg: Make bareiss more flexible
We have several implementations of the Bareiss fraction-free row-reduction
algorithm in the Julia ecosystem. The one in Base was added in #40868
to compute exact determinants. We also have implementations in MTK [1]
and Modia [2].

The MTK and Modia versions additionally have support for custom pivot
selection, open-code a sparse matrix data structure adapted to their
domains and support rank-deficient matrices.

I would like to separate out the algorithmic and data-structures concerns
so that they may tested independently. Of course this function isn't
particularly large, but implementing it correctly and performantly
is still not trivial, so it seems better to have one implementation
rather than three.

[1] https://github.com/SciML/ModelingToolkit.jl/blob/master/src/systems/alias_elimination.jl#L236
[2] https://github.com/ModiaSim/ModiaBase.jl/blob/6c341eed72d9867553cb9565330d8ae85221b343/src/LinearIntegerEquations.jl#L204